BUG: Packages with identical names in virtual environments will not be overwritten.
